media: (modified) change process_time to float

+++ libraries/mediauploader/src/main/java/com/tokopedia/mediauploader/data/entity/UploadHeader.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
 
 data class UploadHeader(
-        @Expose @SerializedName("process_time") val processTime: Int = 0,
+        @Expose @SerializedName("process_time") val processTime: Float = 0f,
         @Expose @SerializedName("reason") val reason: String = "",
         @Expose @SerializedName("error_code") val errorCode: String = ""
 )
